chore: move scroll anchor logic to comman week component
