+++ frontend/packages/app/src/pages/timesheet/project/projectTimesheetTable.tsx
@@ -52,7 +52,7 @@ const ProjectTimesheetGrid = () => {
               return (
                 <Fragment key={`${week.start_date}-${week.end_date}`}>
                   {index === 0 ? (
-                    <div className="sticky top-0 z-10 mb-4 bg-surface-white">
+                    <div className="sticky top-0 z-20 bg-surface-white">
                       <HeaderRow
                         dates={week.dates}
                         showHeading={true}
@@ -73,7 +73,7 @@ const ProjectTimesheetGrid = () => {
                     </div>
                   ) : null}
 
-                  <div className="animate-fade-in">
+                  <div className={cn("animate-fade-in", index === 0 && "mt-4")}>
                     <ProjectTimesheetRow
                       label={week.label}
                       dates={week.dates}
