+++ frontend/packages/app/src/pages/timesheet/team/teamTimesheetWeek.tsx
@@ -13,8 +13,6 @@ import { useTeamTimesheet } from "./context";
 import type { TeamWeekSummary } from "./types";
 import { useTeamWeekMembers } from "./useTeamWeekMembers";
 
-const STICKY_HEADER_OFFSET = 28;
-
 type TeamTimesheetWeekProps = {
   week: TeamWeekSummary;
   defaultExpanded: boolean;
@@ -38,7 +36,6 @@ export const TeamTimesheetWeek = ({
 
   const [expanded, setExpanded] = useState(defaultExpanded);
   const activatedRef = useRef(defaultExpanded);
-  const weekRef = useRef<HTMLDivElement>(null);
 
   if (expanded) {
     activatedRef.current = true;
@@ -71,38 +68,21 @@ export const TeamTimesheetWeek = ({
     root: scrollRoot,
   });
 
-  const handleCollapsedChange = (collapsed: boolean) => {
-    setExpanded(!collapsed);
-    if (!collapsed) return;
-
-    const weekEl = weekRef.current;
-    if (!weekEl) return;
-
-    // Scroll the week into view if it is above the scroll root's top position.
-    const rootTop = scrollRoot?.getBoundingClientRect().top ?? 0;
-    const weekTop = weekEl.getBoundingClientRect().top;
-    if (weekTop < rootTop + STICKY_HEADER_OFFSET) {
-      weekEl.scrollIntoView({ block: "start" });
-    }
-  };
-
   return (
-    <div ref={weekRef} className="scroll-mt-7">
-      <TeamTimesheetRow
-        label={week.label}
-        dates={week.dates}
-        collapsed={!defaultExpanded}
-        onCollapsedChange={handleCollapsedChange}
-        approvalPendingCount={week.approval_pending_count}
-        teamMembers={members}
-        hasMoreMembers={!isFilterRequest && hasMore}
-        isLoadingMembers={
-          !isFilterRequest && (isLoadingMembers || isNextPageLoading)
-        }
-        loadMoreRef={loadMoreRef}
-        setSelectedTask={setSelectedTask}
-        openWeeklyApproval={openWeeklyApproval}
-      />
-    </div>
+    <TeamTimesheetRow
+      label={week.label}
+      dates={week.dates}
+      collapsed={!defaultExpanded}
+      onCollapsedChange={(collapsed) => setExpanded(!collapsed)}
+      approvalPendingCount={week.approval_pending_count}
+      teamMembers={members}
+      hasMoreMembers={!isFilterRequest && hasMore}
+      isLoadingMembers={
+        !isFilterRequest && (isLoadingMembers || isNextPageLoading)
+      }
+      loadMoreRef={loadMoreRef}
+      setSelectedTask={setSelectedTask}
+      openWeeklyApproval={openWeeklyApproval}
+    />
   );
 };
